moved post to routers and added the update post functionality
